requirements.txt
Flask==2.1.0
flask-wtf==0.14.3
requests==2.31.0
gunicorn==20.1.0
firebase-admin==6.2.0
google-cloud-firestore==2.11.1